refactored withdraw to deal with concurrency issues
withdraw is separated into 2 functions now. one to check status of
maiden, and one to handle putting the next player in front of the list.
this allows us to get input from user without the server continuing game
on us. Also, we now need to handle the win case on the next thread
becoming active player.